requirements.txt
# Requirements — Burden of Injuries in India Study
# Python >= 3.11 required

# Core data science
pandas>=2.0.0
numpy>=1.25.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Geospatial (for choropleth maps)
geopandas>=0.14.0
shapely>=2.0.0

# Excel and Office formats
openpyxl>=3.1.0
xlrd>=2.0.1

# PDF table extraction (try both; one may work better for specific PDFs)
tabula-py>=2.7.0
# camelot-py[cv]>=0.11.0  # uncomment if tabula-py fails (requires opencv)

# Parquet support
pyarrow>=12.0.0

# Optional: for advanced geospatial
# fiona>=1.9.0
# pyproj>=3.5.0

# Optional: for R integration
# rpy2>=3.5.0

# Reporting
jinja2>=3.1.0

# Utilities
pathlib2>=2.3.0
python-dateutil>=2.8.0
